wip: derive OptionableConvert for derived optioned types
likely a dead end as we would need to require that T:Optioned
is an Optionable that resolved to itself. While plausible this
is not encoded in the current trait system and likely would need it's own marker trait.

(some OptionableSelf that auto-implements Optionable so that we can restrict
T:Optioned to be OptionableSelf...)