feat!: std and alloc features to support opting-out of linking to the stdlib
The `Error` struct has a breaking change as we had to go from `missing_fields` to `missing_field` to avoid
depending on `Vec`. It also reflects the current codegen situation better. We might still want to bring
`missing_fields` back (together with an Error num) later on (limited to `std`/`alloc`)