feat!: rework kube derive tooling
[kube PR 1850](https://github.com/kube-rs/kube/pull/1850) added support to add attributes for the `CustomResource` root type.

This way we could remove a lot of the user facing specialized tooling (some remained in the codegen to handle the special adjustments
required for the root type).

All attribute and derive macros in the `optionable::kube` package are removed. The `example/k8s` folder documents how to setup ApplyConfiguration (optioned types)
for custom resource definitions. The reworked aproach allows granular conctrol over the derives without railroading the user into a
fixed set of derives for their optioned types.