feat: full support unsized generic types for convert
The previous implementation derived trait bounds for the `OptionableConvert` implementation requiring
`T: Sized + ...`
for any generic parameter `T` used outside of a `required` field.

This was overly strict as it required generic types that are always covered by sized wrappers like `Box` or `Cow` to be stil sized.
The adjusted implementation scans the fields for the actual usage of generics and requires type bounds for them.
I.e. if we have a
```rust
struct A<T>{
  field: Cow<'a, T>
}
```
the corresponding type bound for `OptionableConvert` in the old implementation would have been `T: Sized + ...` while now
it has become `Cow<'a, T>: Sized + ...`.