fix!: no longer generate an optioned type for `WatchEvent` from `k8s-openapi`
WatchEvent is a very special case as it is an externally tagged enum where two variants
share a tag. There is no obvious way how to model this without using a custom serialization/deserialization.
As a WatchEvent is usually only consumed and not patched there is likely not a use case for having
an optioend WatchEvent so it is dropped for now.