fix: move `api_version` and `kind` fields for kubernetes resource types to the front
Having them at the front follows the general convention around these. Another advantage is that they are put out in the conventional
order of a gvk api envelope (even though order does not matter for yaml dict fields it's helpful for human readers).
Also the generated deserialization code first checks the envelope fields this way which is more efficient than starting with filling in metadata,
spec and status prior to checking the envelope.

One downside is that now two `PhantomData` fields are at the front of the resource type definition, but if filling missing fields up
with `..Default::default()` this doesn't really matter.