feat!: replace all dependency-features with versioned variants
Due to the orphanrule we have to have implementations for third party crates in the main `optionable`-crate.
This means that a breaking version upgrade e.g. in `k8s-openapi` (which has many breaking changes as it follows the upstream Kubernetes release schedule)
would cause a breaking change for optionable itself. This would be long-term quite bad if users happen to adopt it but
the trait itself constantly undergoes breaking changes.

We also cannot use the semver-trick (re-exporting e.g. v0.23 in the last release of v0.22 if nothing changed for e.g. the trait definition) as then
the orphanrule would prevent us to add implementations.

Therefore all implementations for third-party crates are now behind versioned features utilizing the cargo package rename functionality.
This allows to e.g. have implementations for `k8s-openapi v0.26` and `k8s-openapi v0.27` feature-gated without undergoing a breaking change.

We will have the `kube v3` breaking change in the future which will not be covered this way (as kube v3 includes new API to set attributes for the root type
of a CRD which we already utilized on the `kube` branch to improve our own API a lot). So we will drop support for `kube v2` once v3 is out with a breaking change.