feat: add type-level `option-wrap` derive argument
Enables to have an already optioned field `Option<T>` become an `Option<Option<T>>` for the optioned type.

While not useful once serialization is involved this setup can be useful for struct patching if no serialization is involved in-between.
This way one can e.g. unset a field by setting it to `Some(None)` on the optioned type.